# openapiupdated.yaml
openapi: 3.0.3
info:
  title: PokerMaster Arena API
  version: "1.0.0"
  description: >
    REST API for PokerMaster Arena, an online Texas Hold’em platform.
    This API supports user authentication, profile management, game room creation/joining,
    in-game actions, spectator mode, game analytics, AI probability calculations, friend management,
    and user preferences.
servers:
  - url: https://api.pokermasterarena.com/v1
paths:
...
  /games:
    post:
      summary: Create a game room
      description: >
        Create a new game room.
        For private games a password is required.
        The creator specifies blind levels, starting chips, and maximum players (2-10).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameCreationRequest'
      responses:
        '201':
          description: Game room created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: Invalid request parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List public game rooms
      description: Retrieve all available public game rooms.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of public game rooms.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /games/{gameId}:
    get:
      summary: Get game details
      description: Retrieve detailed information about a specific game room.
      security:
        - bearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Game details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                
    delete:
      summary: Game room is deleted.
      description: A game room is deleted, because it has fewer than 2 players.
      security:
        - bearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Game deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Game not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /games/{gameId}/join:
    post:
      summary: Join a game room
      description: Join a game room by providing the game ID and password.
      security:
        - bearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
              required:
                - password
      responses:
        '200':
          description: Successfully joined the game room.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Incorrect password or invalid game state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                
    delete:
      summary: A player leaves the game room.
      description: A player leaves the game room of the given game room id. If there remain more than two players in the room, the game may proceed, otherwise the game room is deleted.
      security:
        - bearerAuth: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Player deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Game not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'                      
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    Player:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        gameId:
          type: integer
        credit:
          type: integer
        hand:
          type: array
          items:
            type: string
...
    GameCreationRequest:
      type: object
      properties:
        isPublic:
          type: boolean
          description: Indicates if the game room is private.
        password:
          type: string
          description: Required if isPublic is false.
        smallBlind:
          type: number
        bigBlind:
          type: number
        startCredit:
          type: number
        maximalPlayers:
          type: integer
          minimum: 2
          maximum: 10
      required:
        - isPublic
        - startCredit
        - maximalPlayers
    Game:
      type: object
      properties:
        id:
          type: integer
        isPublic:
          type: boolean
        smallBlind:
          type: number
        bigBlind:
          type: number
        smalllBlindIndex:
          type: integer
        startCredit:
          type: number
        maximalPlayers:
          type: integer
        pot:
          type: integer
        callAmount:
          type: integer
        players:
          type: array
          items:
            $ref: '#/components/schemas/Player'
        status:
          type: string
          enum: [waiting, in-progress, finished]
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []